separate diet entry from original screen
